+++ modules/@apostrophecms/schema/ui/apos/components/AposArrayEditor.vue
@@ -121,6 +121,7 @@ export default {
         type: 'overlay',
         showModal: false
       },
+      titleFieldChoices: null,
       // If we don't clone, then we're making
       // permanent modifications whether the user
       // clicks save or not
@@ -210,6 +211,8 @@ export default {
       await this.nextTick();
       aposSchema.scrollFieldIntoView(name);
     }
+    this.titleFieldChoices = await this.getTitleFieldChoices();
+
   },
   methods: {
     async select(_id) {
@@ -354,7 +357,19 @@ export default {
     label(item) {
       let candidate;
       if (this.field.titleField) {
+
+        // Initial field value
         candidate = get(item, this.field.titleField);
+
+        // If the titleField references a select input, use the
+        // select label as the slat label, rather than the value.
+        if (this.titleFieldChoices) {
+          const choice = this.titleFieldChoices.find(choice => choice.value === candidate);
+          if (choice && choice.label) {
+            candidate = choice.label;
+          }
+        }
+
       } else if (this.schema.find(field => field.name === 'title') && (item.title !== undefined)) {
         candidate = item.title;
       }
@@ -374,6 +389,42 @@ export default {
         title: this.label(item)
       }));
       return result;
+    },
+    async getTitleFieldChoices() {
+      // If the titleField references a select input, get it's choices
+      // to use as labels for the slat UI
+
+      let choices = null;
+      const titleField = this.schema.find(field => field.name === this.field.titleField);
+
+      // The titleField is a select
+      if (titleField?.choices) {
+
+        // Choices are provided by a method
+        if (typeof titleField.choices === 'string') {
+          const action = `${this.moduleOptions.action}/choices`;
+          try {
+            const result = await apos.http.get(
+              action,
+              {
+                qs: {
+                  fieldId: titleField._id
+                }
+              }
+            );
+            if (result && result.choices) {
+              choices = result.choices;
+            }
+          } catch (e) {
+            console.error(this.$t('apostrophe:errorFetchingTitleFieldChoicesByMethod', { name: titleField.name }));
+          }
+
+        // Choices are a normal, hardcoded array
+        } else if (Array.isArray(titleField.choices)) {
+          choices = titleField.choices;
+        }
+      }
+      return choices;
     }
   }
 };