+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 ### Fixes
 
 * Fixes login maximum attempts error message that wasn't showing the plural when lockoutMinutes is more than 1.
+* Fixes the text color of the current array item's slat label in the array editor modal 
+* Fixes the maximum width of an array item's slat label so as to not obscure the Remove button in narrow viewports
+
+### Changes
+
+* If an array field's titleField option is set to a select field, use the selected option's label as the slat label rather it's value.
 * Disable the slat controls of the attachment component while uploading.
 * Fixes bug when re-attaching the same file won't trigger an upload.
 * AposSlat now fully respects the disabled state.